requirements.txt
streamlit==1.28.1
gspread==6.2.1
google-auth==2.47.0
google-auth-oauthlib==1.3.0
pandas==2.0.3